Add files via upload
The database is about an Online Charity Organization, people or companies are able to donate money and to different campaigns. Money is used to help any type of beneficiary as long as a campaign backs them, it also keeps track of who is getting any money. The Charirty organization has staff that has their data stored in it and what campaigns are supported.